feat(AdminIndexView.vue): add form to update holiday settings
Add a form to update holiday settings in the admin dashboard. The form includes a checkbox to control the visibility of the holiday label on the site, and a date picker to select the date of the holiday. When the form is submitted, the holiday settings are updated in the database and a success notification is displayed. If there is an error, an error notification is displayed.