feat(App.vue, VBanner.vue, booking/index.js, settings.js): add support for fetching and updating general settings
This commit adds support for fetching and updating general settings. The `BookingService` class now has a `generalRef` property that references the "general" document in the "booking" collection. The `getGeneralSettings` method fetches the data from this document. The `updateSettings` method updates the "general" document with the provided values.

The `useSettingsStore` store has been created to manage the state of the general settings. It has a `data` property that holds the fetched data, a `loading` property that indicates whether the data is being fetched, an `error` property that indicates whether an error occurred while fetching the data, and a `holidayClosed` property that indicates whether the holiday banner should be closed. The `fetch` method fetches the data from the "general" document and updates the `data` property. The `closeHolidayBanner` method sets the `holidayClosed` property to true.

The `VBanner` component has been added to display a banner when the holiday is active. The banner is displayed only if the `holidayClosed` property is false, the data has been fetched, and the holiday is active. The banner can be dismissed by clicking the "X" button.

The `App.vue` component now fetches the general settings on mount.